Organize into lib/minitest_to_rspec/input
