--- lib/minitest_to_rspec/model/base.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../sexp_assertions'
-
-module MinitestToRspec
-  module Model
-    # Classes inheriting `Base` are simple data objects
-    # representing the S-expressions they are named after.
-    class Base
-      include SexpAssertions
-    end
-  end
-end
